Rename `message-types.ts` in `viz-expr` package to `vscode-and-webview-messages.ts`
